Initial Pass on .Extensions with tests passing.
